requirements.txt
adal==1.2.7
alembic==1.9.4
asgiref==3.6.0
asttokens==2.2.1
azure==4.0.0
backcall==0.2.0
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==39.0.1
debugpy==1.6.6
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.2.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==2.8
inflection==0.5.1
ipython==7.34.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
msrest==0.7.1
msrestazure==0.6.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
oauthlib==3.2.2
openpyxl==3.0.10
packaging==23.0
parso==0.8.3
phonenumbers==8.13.6
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.37
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pymssql==2.2.7
pyodbc==4.0.35
python-dateutil==2.8.2
pyzmq==25.0.0
requests==2.21.0
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.4
sqlalchemy-orm==1.2.4
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.24.3
wcwidth==0.2.6
Werkzeug==2.2.2